fix: use explicit double casts in crofton_rand01 for type safety
Agent-Logs-Url: https://github.com/starseeker/brlcad_quickiterate/sessions/35272e3f-00f8-4d06-8d55-36f2d659b879

+++ brlcad/src/librt/primitives/crofton.cpp
@@ -180,7 +180,7 @@ crofton_worker(int id, void *data)
 static double
 crofton_rand01(void)
 {
-    return rand() / (RAND_MAX + 1.0);
+    return (double)rand() / ((double)RAND_MAX + 1.0);
 }
